+++ docs/conf.py
@@ -28,5 +28,5 @@ html_static_path = []
 exclude_patterns = []
 
 intersphinx_mapping = {
-    'python': ('http://docs.python.org/', None),
+    'python': ('http://docs.python.org/3/', None),
 }