+++ ietfparse/headers.py
@@ -1,0 +1,87 @@
+"""
+Functions for parsing headers.
+
+- :func:`.parse_content_type`: parse a ``Content-Type`` value
+- :func:`.parse_http_accept_header`: parse an ``Accept`` style header
+
+This module also defines classes that might be of some use outside
+of the module.  They are not designed for direct usage unless otherwise
+mentioned.
+
+"""
+
+import functools
+import re
+
+from . import datastructures
+
+
+_COMMENT_RE = re.compile(r'\(.*\)')
+
+
+def _remove_comments(value):
+    """:rtype: str"""  # makes PyCharm happy
+    return _COMMENT_RE.sub('', value)
+
+
+def parse_content_type(content_type, normalize_parameter_values=True):
+    """Parse a content type like header.
+
+    :param str content_type: the string to parse as a content type
+    :param bool normalize_parameter_values:
+        setting this to ``False`` will enable strict RFC2045 compliance
+        in which content parameter values are case preserving.
+    :return: a :class:`~ietfparse.datastructures.ContentType` instance
+
+    """
+    parts = _remove_comments(content_type).split(';')
+    content_type, content_subtype = parts.pop(0).split('/')
+    parameters = {}
+    for type_parameter in parts:
+        name, value = type_parameter.split('=')
+        if normalize_parameter_values:
+            value = value.lower()
+        parameters[name.strip()] = value.strip('"').strip()
+
+    return datastructures.ContentType(content_type, content_subtype,
+                                      parameters)
+
+
+def parse_http_accept_header(header_value):
+    """Parse an HTTP accept-like header.
+
+    :param str header_value: the header value to parse
+    :return: a :class:`list` of :class:`.ContentType` instances
+        in decreasing quality order.  Each instance is augmented
+        with the associated quality as a ``float`` property
+        named ``quality``.
+
+    ``Accept`` is a class of headers that contain a list of values
+    and an associated preference value.  The ever present `Accept`_
+    header is a perfect example.  It is a list of content types and
+    an optional parameter named ``q`` that indicates the relative
+    weight of a particular type.  The most basic example is::
+
+        Accept: audio/*;q=0.2, audio/basic
+
+    Which states that I prefer the ``audio/basic`` content type
+    but will accept other ``audio`` sub-types with an 80% mark down.
+
+    .. _Accept: http://tools.ietf.org/html/rfc7231#section-5.3.2
+
+    """
+    headers = [parse_content_type(header)
+               for header in header_value.split(',')]
+    for header in headers:
+        header.quality = float(header.parameters.pop('q', 1.0))
+
+    def ordering(left, right):
+        if left.quality != right.quality:
+            return right.quality - left.quality
+        if left == right:
+            return 0
+        if left > right:
+            return -1
+        return 1
+
+    return sorted(headers, key=functools.cmp_to_key(ordering))