+++ docs/changelog.rst
@@ -3,4 +3,10 @@ Changelog
 
 * Next Release
 
-  - Create something amazing
+  - Initial implementation containing the following functionality:
+      - ``algorithms.select_content_type``
+      - ``datastructures.ContentType``
+      - ``errors.NoMatch``
+      - ``errors.RootException``
+      - ``headers.parse_content_type``
+      - ``headers.parse_http_accept_header``